Refactor table call function
