+++ src/function/table/CMakeLists.txt
@@ -1,8 +1,10 @@
-add_library(kuzu_table_function
+add_subdirectory(call)
+
+add_library(kuzu_table
         OBJECT
         call_functions.cpp
         scan_functions.cpp)
 
 set(ALL_OBJECT_FILES
-        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_table_function>
+        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_table>
         PARENT_SCOPE)
