+++ src/function/table/call/table_info.cpp
@@ -1,0 +1,102 @@
+#include "catalog/catalog.h"
+#include "catalog/catalog_entry/node_table_catalog_entry.h"
+#include "common/keyword/rdf_keyword.h"
+#include "function/table/bind_input.h"
+#include "function/table/call_functions.h"
+
+using namespace kuzu::catalog;
+using namespace kuzu::common;
+
+namespace kuzu {
+namespace function {
+
+struct TableInfoBindData : public CallTableFuncBindData {
+    TableCatalogEntry* tableEntry;
+
+    TableInfoBindData(TableCatalogEntry* tableEntry, std::vector<LogicalType> returnTypes,
+        std::vector<std::string> returnColumnNames, offset_t maxOffset)
+        : CallTableFuncBindData{std::move(returnTypes), std::move(returnColumnNames), maxOffset},
+          tableEntry{tableEntry} {}
+
+    inline std::unique_ptr<TableFuncBindData> copy() const override {
+        return std::make_unique<TableInfoBindData>(tableEntry, columnTypes, columnNames, maxOffset);
+    }
+};
+
+static void tableFunc(TableFunctionInput& input, common::DataChunk& outputChunk) {
+    auto sharedState =
+        ku_dynamic_cast<TableFuncSharedState*, CallFuncSharedState*>(input.sharedState);
+    auto morsel = sharedState->getMorsel();
+    if (!morsel.hasMoreToOutput()) {
+        outputChunk.getValueVector(0)->state->selVector->selectedSize = 0;
+        return;
+    }
+    auto bindData =
+        ku_dynamic_cast<function::TableFuncBindData*, function::TableInfoBindData*>(input.bindData);
+    auto tableEntry = bindData->tableEntry;
+    auto numPropertiesToOutput = morsel.endOffset - morsel.startOffset;
+    auto vectorPos = 0;
+    for (auto i = 0u; i < numPropertiesToOutput; i++) {
+        auto property = &tableEntry->getPropertiesRef()[morsel.startOffset + i];
+        if (tableEntry->getTableType() == TableType::REL) {
+            if (property->getName() == InternalKeyword::ID) {
+                // Skip internal id column.
+                continue;
+            }
+            if (property->getName() == rdf::PID) {
+                // Replace pid column with (virtual) iri column.
+                outputChunk.getValueVector(0)->setValue<int64_t>(vectorPos, -1);
+                outputChunk.getValueVector(1)->setValue(
+                    vectorPos, std::string(rdf::IRI) + " (Virtual)");
+                outputChunk.getValueVector(2)->setValue(
+                    vectorPos, LogicalType::STRING()->toString());
+                vectorPos++;
+                continue;
+            }
+        }
+        outputChunk.getValueVector(0)->setValue(vectorPos, (int64_t)property->getPropertyID());
+        outputChunk.getValueVector(1)->setValue(vectorPos, property->getName());
+        outputChunk.getValueVector(2)->setValue(vectorPos, property->getDataType()->toString());
+
+        if (tableEntry->getTableType() == TableType::NODE) {
+            auto nodeTableEntry =
+                ku_dynamic_cast<TableCatalogEntry*, NodeTableCatalogEntry*>(tableEntry);
+            auto primaryKeyID = nodeTableEntry->getPrimaryKeyPID();
+            outputChunk.getValueVector(3)->setValue(
+                vectorPos, primaryKeyID == property->getPropertyID());
+        }
+        vectorPos++;
+    }
+    outputChunk.state->selVector->selectedSize = vectorPos;
+}
+
+static std::unique_ptr<TableFuncBindData> bindFunc(
+    main::ClientContext* context, TableFuncBindInput* input) {
+    std::vector<std::string> columnNames;
+    std::vector<LogicalType> columnTypes;
+    auto tableName = input->inputs[0].getValue<std::string>();
+    auto tableID = context->getCatalog()->getTableID(context->getTx(), tableName);
+    auto tableEntry = context->getCatalog()->getTableCatalogEntry(context->getTx(), tableID);
+    columnNames.emplace_back("property id");
+    columnTypes.push_back(*LogicalType::INT64());
+    columnNames.emplace_back("name");
+    columnTypes.push_back(*LogicalType::STRING());
+    columnNames.emplace_back("type");
+    columnTypes.push_back(*LogicalType::STRING());
+    if (tableEntry->getTableType() == TableType::NODE) {
+        columnNames.emplace_back("primary key");
+        columnTypes.push_back(*LogicalType::BOOL());
+    }
+    return std::make_unique<TableInfoBindData>(
+        tableEntry, std::move(columnTypes), std::move(columnNames), tableEntry->getNumProperties());
+}
+
+function_set TableInfoFunction::getFunctionSet() {
+    function_set functionSet;
+    functionSet.push_back(std::make_unique<TableFunction>(TABLE_INFO_FUNC_NAME, tableFunc, bindFunc,
+        initSharedState, initEmptyLocalState, std::vector<LogicalTypeID>{LogicalTypeID::STRING}));
+    return functionSet;
+}
+
+} // namespace function
+} // namespace kuzu
