+++ src/include/common/enums/expression_type.h
@@ -236,6 +236,7 @@ const char* const DB_VERSION_FUNC_NAME = "DB_VERSION";
 const char* const CURRENT_SETTING_FUNC_NAME = "CURRENT_SETTING";
 const char* const SHOW_TABLES_FUNC_NAME = "SHOW_TABLES";
 const char* const SHOW_CONNECTION_FUNC_NAME = "SHOW_CONNECTION";
+const char* const STORAGE_INFO_FUNC_NAME = "STORAGE_INFO";
 const char* const READ_PARQUET_FUNC_NAME = "READ_PARQUET";
 const char* const READ_NPY_FUNC_NAME = "READ_NPY";
 const char* const READ_CSV_SERIAL_FUNC_NAME = "READ_CSV_SERIAL";
@@ -250,7 +251,6 @@ const char* const IN_MEM_READ_RDF_LITERAL_FUNC_NAME = "IN_MEM_READ_RDF_LITERAL";
 const char* const IN_MEM_READ_RDF_RESOURCE_TRIPLE_FUNC_NAME = "IN_MEM_READ_RDF_RESOURCE_TRIPLE";
 const char* const IN_MEM_READ_RDF_LITERAL_TRIPLE_FUNC_NAME = "IN_MEM_READ_RDF_LITERAL_TRIPLE";
 const char* const READ_PANDAS_FUNC_NAME = "READ_PANDAS";
-const char* const STORAGE_INFO_FUNC_NAME = "STORAGE_INFO";
 
 enum class ExpressionType : uint8_t {
