+++ src/include/function/table/scan_functions.h
