+++ src/include/processor/operator/call/in_query_call.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "function/table/bind_data.h"
 #include "function/table_functions.h"
-#include "function/table_functions/bind_data.h"
 #include "processor/operator/physical_operator.h"
 
 namespace kuzu {
