+++ src/include/processor/operator/persistent/reader/csv/parallel_csv_reader.h
@@ -3,10 +3,10 @@
 #include "base_csv_reader.h"
 #include "common/types/types.h"
 #include "function/scalar_function.h"
+#include "function/table/bind_data.h"
+#include "function/table/bind_input.h"
+#include "function/table/scan_functions.h"
 #include "function/table_functions.h"
-#include "function/table_functions/bind_data.h"
-#include "function/table_functions/bind_input.h"
-#include "function/table_functions/scan_functions.h"
 
 namespace kuzu {
 namespace processor {
@@ -53,9 +53,8 @@ struct ParallelCSVScan {
 
     static void tableFunc(function::TableFunctionInput& input, common::DataChunk& outputChunk);
 
-    static std::unique_ptr<function::TableFuncBindData> bindFunc(main::ClientContext* /*context*/,
-        function::TableFuncBindInput* input, catalog::Catalog* /*catalog*/,
-        storage::StorageManager* /*storageManager*/);
+    static std::unique_ptr<function::TableFuncBindData> bindFunc(
+        main::ClientContext* /*context*/, function::TableFuncBindInput* input);
 
     static std::unique_ptr<function::TableFuncSharedState> initSharedState(
         function::TableFunctionInitInput& input);
