+++ src/include/processor/operator/persistent/reader/csv/serial_csv_reader.h
@@ -2,9 +2,9 @@
 
 #include "base_csv_reader.h"
 #include "function/scalar_function.h"
-#include "function/table_functions/bind_data.h"
-#include "function/table_functions/bind_input.h"
-#include "function/table_functions/scan_functions.h"
+#include "function/table/bind_data.h"
+#include "function/table/bind_input.h"
+#include "function/table/scan_functions.h"
 
 namespace kuzu {
 namespace processor {
@@ -45,9 +45,8 @@ struct SerialCSVScan {
 
     static void tableFunc(function::TableFunctionInput& input, common::DataChunk& outputChunk);
 
-    static std::unique_ptr<function::TableFuncBindData> bindFunc(main::ClientContext* /*context*/,
-        function::TableFuncBindInput* input, catalog::Catalog* /*catalog*/,
-        storage::StorageManager* /*storageManager*/);
+    static std::unique_ptr<function::TableFuncBindData> bindFunc(
+        main::ClientContext* /*context*/, function::TableFuncBindInput* input);
 
     static std::unique_ptr<function::TableFuncSharedState> initSharedState(
         function::TableFunctionInitInput& input);
