+++ src/include/processor/operator/persistent/reader/parquet/parquet_reader.h
@@ -4,9 +4,9 @@
 #include "common/data_chunk/data_chunk.h"
 #include "common/types/types.h"
 #include "function/scalar_function.h"
-#include "function/table_functions/bind_data.h"
-#include "function/table_functions/bind_input.h"
-#include "function/table_functions/scan_functions.h"
+#include "function/table/bind_data.h"
+#include "function/table/bind_input.h"
+#include "function/table/scan_functions.h"
 #include "parquet/parquet_types.h"
 #include "resizable_buffer.h"
 #include "thrift/protocol/TCompactProtocol.h"
@@ -111,9 +111,8 @@ struct ParquetScanFunction {
 
     static void tableFunc(function::TableFunctionInput& input, common::DataChunk& outputChunk);
 
-    static std::unique_ptr<function::TableFuncBindData> bindFunc(main::ClientContext* /*context*/,
-        function::TableFuncBindInput* input, catalog::Catalog* catalog,
-        storage::StorageManager* /*storageManager*/);
+    static std::unique_ptr<function::TableFuncBindData> bindFunc(
+        main::ClientContext* /*context*/, function::TableFuncBindInput* input);
 
     static std::unique_ptr<function::TableFuncSharedState> initSharedState(
         function::TableFunctionInitInput& input);
