+++ src/processor/operator/persistent/reader/csv/parallel_csv_reader.cpp
@@ -158,8 +158,7 @@ void ParallelCSVScan::tableFunc(TableFunctionInput& input, common::DataChunk& ou
 }
 
 std::unique_ptr<function::TableFuncBindData> ParallelCSVScan::bindFunc(
-    main::ClientContext* /*context*/, function::TableFuncBindInput* input,
-    catalog::Catalog* /*catalog*/, storage::StorageManager* /*storageManager*/) {
+    main::ClientContext* /*context*/, function::TableFuncBindInput* input) {
     auto scanInput = reinterpret_cast<function::ScanTableFuncBindInput*>(input);
     std::vector<std::string> detectedColumnNames;
     std::vector<common::LogicalType> detectedColumnTypes;
