+++ src/processor/operator/persistent/reader/csv/serial_csv_reader.cpp
@@ -80,8 +80,7 @@ void SerialCSVScan::tableFunc(TableFunctionInput& input, DataChunk& outputChunk)
 }
 
 std::unique_ptr<function::TableFuncBindData> SerialCSVScan::bindFunc(
-    main::ClientContext* /*context*/, function::TableFuncBindInput* input,
-    catalog::Catalog* /*catalog*/, storage::StorageManager* /*storageManager*/) {
+    main::ClientContext* /*context*/, function::TableFuncBindInput* input) {
     auto scanInput = reinterpret_cast<function::ScanTableFuncBindInput*>(input);
     std::vector<std::string> detectedColumnNames;
     std::vector<common::LogicalType> detectedColumnTypes;
