+++ src/processor/operator/persistent/reader/npy/npy_reader.cpp
@@ -252,8 +252,7 @@ void NpyScanFunction::tableFunc(TableFunctionInput& input, DataChunk& outputChun
 }
 
 std::unique_ptr<function::TableFuncBindData> NpyScanFunction::bindFunc(
-    main::ClientContext* /*context*/, function::TableFuncBindInput* input,
-    catalog::Catalog* /*catalog*/, storage::StorageManager* /*storageManager*/) {
+    main::ClientContext* /*context*/, function::TableFuncBindInput* input) {
     auto scanInput = reinterpret_cast<function::ScanTableFuncBindInput*>(input);
 
     std::vector<std::string> detectedColumnNames;
